requirements.txt
discord.py==2.3.2
asyncpg==0.30.0
matplotlib==3.9.0
numpy==1.26.4
Pillow==10.4.0
python-dotenv==1.0.1
audioop-lts
